feat(wip): Worked more on bridge testing

+++ contracts/interfaces/IWormhole.sol
@@ -6,7 +6,7 @@ pragma solidity ^0.8.0;
 import "../Structs.sol";
 
 interface IWormhole is Structs {
-    event LogMessagePublished(address indexed sender, uint64 sequence, uint32 nonce, bytes payload, uint8 consistencyLevel);
+    event LogMessagePublished(address indexed sender, uint64 sequence, uint64 nonce, bytes payload, uint8 consistencyLevel);
 
     function publishMessage(
         uint64 nonce,
